Abstracted get distinct category and get category total functions to their relevant util files.
